fix: guard BaoChunk::Leaf size against usize overflow in outboard init
Wire the patched local bao-tree fork (BaoChunk::Leaf.size: usize -> u64).
Cast size to usize with a safe `as usize` in init_impl, which is safe
because this path iterates PostOrderChunkIter whose leaf sizes are bounded
by chunk_group_bytes(). Add [patch.crates-io] to redirect bao-tree to the
local fork that carries the u64 fix.

+++ src/util.rs
@@ -323,7 +323,8 @@ pub(crate) mod outboard_with_progress {
                     if let Err(err) = progress.send(start_chunk).await {
                         return Ok(Err(err));
                     }
-                    let buf = &mut buffer[..size];
+                    // size is bounded by chunk_group_bytes() here, so the cast is safe
+                    let buf = &mut buffer[..size as usize];
                     data.read_exact(buf)?;
                     let hash = hash_subtree(start_chunk.0, buf, is_root);
                     stack.push(hash);
